Made FDPoint it's own class.
